send change to bank when having no free space on backpack

+++ src/Game/NeoServer.Game.Common/Contracts/Creatures/IPlayer.cs
@@ -170,7 +170,7 @@ namespace NeoServer.Server.Model.Players.Contracts
         void StartShopping(IShopperNpc npc);
         void StopShopping();
         bool Sell(IItemType item, byte amount, bool ignoreEquipped);
-        void ReceivePayment(IEnumerable<IItem> coins, uint total);
+        void ReceivePayment(IEnumerable<IItem> coins, ulong total);
         bool CanReceiveInCashPayment(IEnumerable<IItem> coins);
 
         void ReceivePurchasedItems(INpc from, SaleContract saleContract, params IItem[] items);
